pyproject.toml
[tool.poetry]
name = "torrent-downloader"
version = "0.1.0"
description = ""
authors = ["Mikhail Onufrienko <ghast13@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
python-telegram-bot = {extras = ["all"], version = "^21.6"}
alembic = "^1.13.3"
loguru = "^0.7.2"
bencodepy = "^0.9.5"
asyncpg = "^0.30.0"
python-qbittorrent = "^0.4.3"
celery = {extras = ["amqp"], version = "^5.4.0"}
kombu = "^5.4.2"
amqp = "^5.3.1"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
